Polished UI, Kinect "Scroll" and Starting Scene
Changed most of the Buttons sprites, added some new images, added 2 buttons in the bottom left of the screen that become active in the Kinect mode and can be used to scroll the spawned objects list. Added a start menu so that the user can choose what type of interaction he wants to use.