Added Terrain
Will do it so that the whiteboard will have 2 modes: one for ideas and another one for the terrain. The terrain mode, will make the whiteboard green and will let he user paint with 2 colors (blue for water, something else for mountain), and that will dynamically change the terrain so that it has mountains and water where the user drew them on the board.